"final" trait changes
