+++ app/Http/Traits/FilterTrait.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Traits;
 
+use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Http\Request;
@@ -133,5 +134,43 @@ trait FilterTrait
         }
 
         return $rtn;
+    }
+
+    /**
+     * @author EB
+     * @return Carbon[]
+     */
+    protected function getDateRange()
+    {
+        $defaultDates = [
+            'date_to' => Carbon::now(),
+            'date_from' => Carbon::today(),
+        ];
+
+        $filters = $this->getFilters();
+
+        if($filters->has('date_to')) {
+            $defaultDates['date_to'] = Carbon::createFromFormat('Y/m/d', $filters['date_to'])->endOfDay();
+            $filters->forget('date_to');
+        }
+
+        if($filters->has('date_from')) {
+            $defaultDates['date_from'] = Carbon::createFromFormat('Y/m/d', $filters['date_from'])->startOfDay();
+            $filters->forget('date_from');
+        }
+
+        return $defaultDates;
+    }
+
+    /**
+     * @param Builder $model
+     * @param string $field
+     * @param Carbon $after
+     * @param Carbon $before
+     * @return Builder
+     */
+    protected function processDateFilters(Builder $model, $field, Carbon $after, Carbon $before)
+    {
+        return $model->where($field, '>', $after->toDateTimeString())->where($field, '<', $before->toDateTimeString());
     }
 }
